requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
psycopg2-binary>=2.9
python-dotenv>=1.0
django-environ>=0.10.0
pytest>=8.2
pytest-django>=4.7
requests>=2.32
gunicorn>=21.2
django-jazzmin>=2.6,<3.0 